Implement more (and clarify existing) static Migration methods. Use the new methods in status and verify. Bump version number.
--HG--
extra : convert_revision : svn%3A4fd3d1b2-7acc-2142-88d6-33460400c3cb/ladder/trunk%409084

+++ system/core/commands/verify.php
@@ -1,17 +1,26 @@
 <?php
 
 // Get all the migrations from the filesystem.
-$migrations = Migration::get_migrations();
+$migrations = Migration::get_migration_ids();
 
 // Get a Database instance.
 $db = Database::factory();
 
 // Loop over each database.
 while ($db->next_database()) {
+	// Get the migrations from this database.
 	$db_migrations = $db->get_migrations();
 
-	foreach (array_diff($db_migrations, $migrations) as $db_id) {
-		echo 'Removing invalid migration ', $db_id, "\n";
-		$db->remove_migration($db_id);
+	// Find any that are invalid.
+	$bad_ids = array_diff($db_migrations, $migrations);
+
+	// Remove any bad ones, or just output 'Verified' if none.
+	if ((bool) $bad_ids) {
+		foreach ($bad_ids as $db_id) {
+			echo 'Removing invalid migration ', $db_id, "\n";
+			$db->remove_migration($db_id);
+		}
+	} else {
+		echo "\t", 'Verified', "\n";
 	}
 }
